moved more state inside
